+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/replacedSessionCookieDetection/DetectionCookieValueTest.java
@@ -11,7 +11,7 @@ public class DetectionCookieValueTest {
     @Test
     public void test_isWithInLimit_default() {
         // given
-        final DetectionCookieValue cookieValue = new DetectionCookieValue();
+        final DetectionCookieValue cookieValue = new DetectionCookieValue(System.currentTimeMillis(), 1 * 60 * 1000);
         // when
         final boolean withInLimit = cookieValue.isWithInLimit();
         // then
@@ -21,7 +21,7 @@ public class DetectionCookieValueTest {
     @Test
     public void test_isWithInLimit_new() {
         // given
-        final DetectionCookieValue cookieValue = new DetectionCookieValue("0:" + System.currentTimeMillis());
+        final DetectionCookieValue cookieValue = new DetectionCookieValue(0 + SPLITTER + System.currentTimeMillis());
         // when
         final boolean withInLimit = cookieValue.isWithInLimit();
         // then
@@ -72,7 +72,7 @@ public class DetectionCookieValueTest {
     @Test
     public void test_contstructor_default() {
         // given
-        final DetectionCookieValue cookieValue = new DetectionCookieValue();
+        final DetectionCookieValue cookieValue = new DetectionCookieValue(System.currentTimeMillis(), 1 * 60 * 1000);
         // when
         // then
         Assertions.assertEquals(0, cookieValue.counter);
@@ -81,11 +81,11 @@ public class DetectionCookieValueTest {
     @Test
     public void test_contstructor_with_one() {
         // given
-        final long accessTime = System.currentTimeMillis();
-        final DetectionCookieValue cookieValue = new DetectionCookieValue(1 + SPLITTER + accessTime);
+        final long lifeTime = System.currentTimeMillis();
+        final DetectionCookieValue cookieValue = new DetectionCookieValue(1 + SPLITTER + lifeTime);
         // when
         // then
         Assertions.assertEquals(1, cookieValue.counter);
-        Assertions.assertEquals(accessTime, cookieValue.lastSessionAccess);
+        Assertions.assertEquals(lifeTime, cookieValue.sessionLifeTime);
     }
 }