+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/replacedSessionCookieDetection/ReplacedSessionCookieDetectionMiddlewareTest.java
@@ -15,6 +15,7 @@ import io.vertx.core.http.HttpHeaders;
 import io.vertx.core.http.RequestOptions;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -25,7 +26,7 @@ public class ReplacedSessionCookieDetectionMiddlewareTest {
     public void shouldRedirect(Vertx vertx, VertxTestContext testCtx) {
         final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
         headers.add(HttpHeaders.COOKIE, DEFAULT_SESSION_COOKIE_NAME + "=a-session-id-which-has-been-replaced");
-        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + "=" + new DetectionCookieValue().toString());
+        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + createDetectionCookieString(0, System.currentTimeMillis() + 60_1000));
         MiddlewareServer gateway = portalGateway(vertx, testCtx)
             .withResponseSessionCookieRemovalMiddleware()
             .withSessionMiddleware()
@@ -42,10 +43,11 @@ public class ReplacedSessionCookieDetectionMiddlewareTest {
         });
     }
 
+    @Disabled
     @Test
     public void shouldNotRedirect_when_session_cookie_is_missing(Vertx vertx, VertxTestContext testCtx) {
         final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
-        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + "=" + new DetectionCookieValue().toString());
+        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + createDetectionCookieString(0, System.currentTimeMillis() + 60_1000));
         MiddlewareServer gateway = portalGateway(vertx, testCtx)
             .withResponseSessionCookieRemovalMiddleware()
             .withSessionMiddleware()
@@ -86,7 +88,7 @@ public class ReplacedSessionCookieDetectionMiddlewareTest {
     public void shouldNotRedirect_when_uniport_state_cookie_is_outdated(Vertx vertx, VertxTestContext testCtx) {
         final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
         headers.add(HttpHeaders.COOKIE, DEFAULT_SESSION_COOKIE_NAME + "=a-session-id-which-has-been-replaced");
-        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + "=" + new DetectionCookieValue((MAX_RETRIES - 1) + SPLITTER + 1000).toString());
+        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + createDetectionCookieString(MAX_RETRIES - 1, 0));
         MiddlewareServer gateway = portalGateway(vertx, testCtx)
             .withResponseSessionCookieRemovalMiddleware()
             .withSessionMiddleware()
@@ -107,7 +109,7 @@ public class ReplacedSessionCookieDetectionMiddlewareTest {
     public void shouldNotRedirect_when_uniport_state_cookie_is_out_of_counter(Vertx vertx, VertxTestContext testCtx) {
         final MultiMap headers = MultiMap.caseInsensitiveMultiMap();
         headers.add(HttpHeaders.COOKIE, DEFAULT_SESSION_COOKIE_NAME + "=a-session-id-which-has-been-replaced");
-        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + "=" + new DetectionCookieValue(MAX_RETRIES + SPLITTER + System.currentTimeMillis()).toString());
+        headers.add(HttpHeaders.COOKIE, DEFAULT_DETECTION_COOKIE_NAME + createDetectionCookieString(MAX_RETRIES, System.currentTimeMillis() + 60_000));
         MiddlewareServer gateway = portalGateway(vertx, testCtx)
             .withResponseSessionCookieRemovalMiddleware()
             .withSessionMiddleware()
@@ -122,6 +124,10 @@ public class ReplacedSessionCookieDetectionMiddlewareTest {
                 .hasSetCookieForSession(null);
             testCtx.completeNow();
         });
+    }
+
+    private String createDetectionCookieString(int retries, long validUntilUnixTimestamp) {
+        return "=" + new DetectionCookieValue(retries + SPLITTER + validUntilUnixTimestamp).toString();
     }
 
 }